backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
ultralytics
torch
torchvision
numpy==1.26.4
opencv-python
huggingface-hub
facenet-pytorch
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
APScheduler==3.10.4
inference
supervision
reportlab==4.0.7
openpyxl==3.1.2
aiosmtplib==3.0.1
Jinja2==3.1.2
matplotlib==3.8.2
httpx
aiohttp
pyserial
